+++ packages/contracts/contracts/test/unit/DepositManager.t.sol
@@ -17,8 +17,6 @@ import {Teller} from "../../Teller.sol";
 import {TestJoinSplitVerifier} from "../harnesses/TestJoinSplitVerifier.sol";
 import {TestSubtreeUpdateVerifier} from "../harnesses/TestSubtreeUpdateVerifier.sol";
 import {SimpleERC20Token} from "../tokens/SimpleERC20Token.sol";
-import {SimpleERC721Token} from "../tokens/SimpleERC721Token.sol";
-import {SimpleERC1155Token} from "../tokens/SimpleERC1155Token.sol";
 import {WETH9} from "../tokens/WETH9.sol";
 
 contract DepositManagerTest is Test {
@@ -28,8 +26,6 @@ contract DepositManagerTest is Test {
     WETH9 public weth;
 
     SimpleERC20Token[3] ERC20s;
-    SimpleERC721Token[3] ERC721s;
-    SimpleERC1155Token[3] ERC1155s;
 
     string constant CONTRACT_NAME = "NocturneDepositManager";
     string constant CONTRACT_VERSION = "v1";
@@ -106,19 +102,9 @@ contract DepositManagerTest is Test {
         // Instantiate token contracts
         for (uint256 i = 0; i < 3; i++) {
             ERC20s[i] = new SimpleERC20Token();
-            ERC721s[i] = new SimpleERC721Token();
-            ERC1155s[i] = new SimpleERC1155Token();
 
             handler.setSupportedContractAllowlistPermission(
                 address(ERC20s[i]),
-                true
-            );
-            handler.setSupportedContractAllowlistPermission(
-                address(ERC721s[i]),
-                true
-            );
-            handler.setSupportedContractAllowlistPermission(
-                address(ERC1155s[i]),
                 true
             );